removed TCO policy resource
