--- coralogix/data_source_coralogix_tco_policy_logs.go
@@ -1,110 +1,0 @@
-// Copyright 2024 Coralogix Ltd.
-// 
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-// 
-//     https://www.apache.org/licenses/LICENSE-2.0
-// 
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package coralogix
-
-import (
-	"context"
-	"fmt"
-	"log"
-
-	"terraform-provider-coralogix/coralogix/clientset"
-	tcopolicies "terraform-provider-coralogix/coralogix/clientset/grpc/tco-policies"
-
-	"google.golang.org/protobuf/encoding/protojson"
-
-	"github.com/hashicorp/terraform-plugin-framework/datasource"
-	"github.com/hashicorp/terraform-plugin-framework/resource"
-	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/status"
-
-	"google.golang.org/protobuf/types/known/wrapperspb"
-)
-
-var _ datasource.DataSourceWithConfigure = &TCOPolicyDataSource{}
-
-func NewTCOPolicyDataSource() datasource.DataSource {
-	return &TCOPolicyDataSource{}
-}
-
-type TCOPolicyDataSource struct {
-	client *clientset.TCOPoliciesClient
-}
-
-func (d *TCOPolicyDataSource) Metadata(_ context.Context, req datasource.MetadataRequest, resp *datasource.MetadataResponse) {
-	resp.TypeName = req.ProviderTypeName + "_tco_policy_logs"
-}
-
-func (d *TCOPolicyDataSource) Configure(_ context.Context, req datasource.ConfigureRequest, resp *datasource.ConfigureResponse) {
-	if req.ProviderData == nil {
-		return
-	}
-
-	clientSet, ok := req.ProviderData.(*clientset.ClientSet)
-	if !ok {
-		resp.Diagnostics.AddError(
-			"Unexpected Resource Configure Type",
-			fmt.Sprintf("Expected *clientset.ClientSet, got: %T. Please report this issue to the provider developers.", req.ProviderData),
-		)
-		return
-	}
-
-	d.client = clientSet.TCOPolicies()
-}
-
-func (d *TCOPolicyDataSource) Schema(ctx context.Context, _ datasource.SchemaRequest, resp *datasource.SchemaResponse) {
-	var r TCOPolicyResource
-	var resourceResp resource.SchemaResponse
-	r.Schema(ctx, resource.SchemaRequest{}, &resourceResp)
-
-	resp.Schema = frameworkDatasourceSchemaFromFrameworkResourceSchema(resourceResp.Schema)
-}
-
-func (d *TCOPolicyDataSource) Read(ctx context.Context, req datasource.ReadRequest, resp *datasource.ReadResponse) {
-	var data *TCOPolicyResourceModel
-	resp.Diagnostics.Append(req.Config.Get(ctx, &data)...)
-	if resp.Diagnostics.HasError() {
-		return
-	}
-
-	//Get refreshed tco-policy value from Coralogix
-	id := data.ID.ValueString()
-	log.Printf("[INFO] Reading tco-policy: %s", id)
-	getPolicyReq := &tcopolicies.GetPolicyRequest{Id: wrapperspb.String(id)}
-	getPolicyResp, err := d.client.GetTCOPolicy(ctx, getPolicyReq)
-	if err != nil {
-		log.Printf("[ERROR] Received error: %s", err.Error())
-		if status.Code(err) == codes.NotFound {
-			resp.Diagnostics.AddWarning(
-				err.Error(),
-				fmt.Sprintf("tco-policy %q is in state, but no longer exists in Coralogix backend", id),
-			)
-		} else {
-			resp.Diagnostics.AddError(
-				"Error reading tco-policy",
-				formatRpcErrors(err, getTCOPolicyURL, protojson.Format(getPolicyReq)),
-			)
-		}
-		return
-	}
-	log.Printf("[INFO] Received tco-policy: %s", protojson.Format(getPolicyResp))
-
-	data, diags := flattenTCOPolicy(ctx, getPolicyResp.GetPolicy())
-	if diags.HasError() {
-		resp.Diagnostics = diags
-		return
-	}
-	// Save data into Terraform state
-	resp.Diagnostics.Append(resp.State.Set(ctx, data)...)
-}
